Final fixes required to work in next
