+++ src/App.css
@@ -1,7 +1,8 @@
 .appContainer {
   display: flex;
   flex-direction: row;
-  margin: 2rem;
+  padding: 1rem;
+  height: 80vh;
 }
 
 .instructions {
